Event latest update with Wedding endpoints and EventEntity relationships

+++ src/main/java/edu/icet/repository/event/EventSummaryRepository.java
@@ -3,5 +3,5 @@ package edu.icet.repository.event;
 import edu.icet.entity.event.EventSummaryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventSummaryRepository extends JpaRepository<EventSummaryEntity, Integer> {
+public interface EventSummaryRepository extends JpaRepository<EventSummaryEntity, Long> {
 }
